2to 3 int long change

+++ webob/datetime_utils.py
@@ -17,7 +17,7 @@ import time
 
 from webob.compat import (
     integer_types,
-    int,
+    long,
     native_,
     text_type,
     )
@@ -110,7 +110,7 @@ def parse_date_delta(value):
 
 
 def serialize_date_delta(value):
-    if isinstance(value, (float, int)):
+    if isinstance(value, (float, long)):
         return str(int(value))
     else:
         return serialize_date(value)
